requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
black==24.4.2
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
contextlib2==21.6.0
cryptography==42.0.8
distlib==0.3.8
Django==5.0.6
django-ninja==1.1.0
django-ninja-extra==0.20.7
filelock==3.15.4
flake8==7.1.0
identify==2.5.36
idna==3.7
injector==0.21.0
isort==5.13.2
jwcrypto==1.5.6
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pre-commit==3.7.1
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
